af_scaletempo2: fix initial output_time value
This fixes negative search window position that could happen during the
first WSOLA iterations. The output time is updated roughly proportional
to playback speed and is used to calculate the new center of the search
block. Initializing it with the `search_block_center_offset` is
consistent with the initial search block position of 0.